refactor(tray): extract CommandCenterStateBuilder from App.xaml.cs
Moves the BuildCommandCenterState cluster (11 methods, ~400 lines) into
two new classes in Services/:

- AppStateSnapshot: sealed record capturing the 15 App fields the cluster
  reads at call time (init-only, no mutation).
- CommandCenterStateBuilder: receives the snapshot, exposes Build() returning
  GatewayCommandCenterState.

App.BuildCommandCenterState() becomes a 2-line delegation call. All 11
existing callers are unchanged.

The cluster is near-pure: reads App state, no shared mutations, no UI
calls, no DispatcherQueue marshal. Equivalence is verifiable by comparing
the DTO output field by field. Execution order including the
ApplyDetectedSshForwardTopology mutation of topology in its fixed position
is preserved exactly.
